+++ backend/src/main/java/org/maklashev/goldengroup/model/repositories/ProductionListRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface ProductionListRepository extends JpaRepository<ProductionList, Long> {
     @Query("SELECT id FROM ProductionList WHERE productionStart >= :startDate AND productionFinish <= :endDate")
-    List<Integer> findIdsInDateRange(@Param("startDate") LocalDateTime startDate, @Param("endDate") LocalDateTime endDate);
+    List<Long> findIdsInDateRange(@Param("startDate") LocalDateTime startDate, @Param("endDate") LocalDateTime endDate);
 
 
 }
